Made filters more modular for code refactor. Added filter parity between languages/articles overview tabs
